T1 backend courses info

+++ coursesModule/src/main/java/com/example/coursesmodule/api/ResourceController.java
@@ -1,25 +1,42 @@
 package com.example.coursesmodule.api;
 
 import com.example.coursesmodule.model.Resource;
+import com.example.coursesmodule.service.ComponentService;
 import com.example.coursesmodule.service.ResourceService;
+import com.example.coursesmodule.service.SubjectService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
+import static org.springframework.http.HttpStatus.*;
+
 @RestController
 @RequestMapping("api/v1/subjects/{subjectTitle}/components/{componentType}/resources")
+@CrossOrigin(origins = "*")
 public class ResourceController {
+    
+    private static final String SUBJECT_ERROR = "Subject not found";
+    private static final String COMPONENT_ERROR = "Component not found";
+    private static final String RESOURCE_ERROR = "Resource not found";
     private final ResourceService resourceService;
-
+    private final SubjectService subjectService;
+    private final ComponentService componentService;
     @Autowired
-    public ResourceController(ResourceService resourceService) {
+    public ResourceController(ResourceService resourceService, SubjectService subjectService, ComponentService componentService) {
         this.resourceService = resourceService;
+        this.subjectService = subjectService;
+        this.componentService = componentService;
     }
 
     @GetMapping
     public List<Resource> getResources(@PathVariable("subjectTitle") String title,
                                        @PathVariable("componentType") String type) {
+        if(subjectService.getSubjectByTitle(title).isEmpty())
+            throw new ResponseStatusException(NOT_FOUND, SUBJECT_ERROR);
+        if(componentService.getComponentByType(title, type).isEmpty())
+            throw new ResponseStatusException(NOT_FOUND, COMPONENT_ERROR);
         return resourceService.getResources(title, type);
     }
 
@@ -27,26 +44,38 @@ public class ResourceController {
     public Resource getResourceByTitle(@PathVariable("subjectTitle") String title,
                                        @PathVariable("componentType") String type,
                                        @PathVariable("resourceTitle") String resourceTitle) {
+        if(subjectService.getSubjectByTitle(title).isEmpty())
+            throw new ResponseStatusException(NOT_FOUND, SUBJECT_ERROR);
+        if(componentService.getComponentByType(title, type).isEmpty())
+            throw new ResponseStatusException(NOT_FOUND, COMPONENT_ERROR);
         return resourceService.getResourceByTitle(title, type, resourceTitle)
-                .orElseThrow(() -> new IllegalStateException("Resource not found"));
+                .orElseThrow(() -> new ResponseStatusException(NOT_FOUND, RESOURCE_ERROR));
     }
 
     @PostMapping
     public void addResource(@PathVariable("subjectTitle") String title,
                             @PathVariable("componentType") String type,
                             @RequestBody Resource resource) {
-        if(resourceService.addResource(title, type, resource) == 0) {
-            throw new IllegalStateException("Invalid title given");
-        }
+        if(subjectService.getSubjectByTitle(title).isEmpty())
+            throw new ResponseStatusException(NOT_FOUND, SUBJECT_ERROR);
+        if(componentService.getComponentByType(title, type).isEmpty())
+            throw new ResponseStatusException(NOT_FOUND, COMPONENT_ERROR);
+        if(resourceService.addResource(title, type, resource) == 0)
+            throw new ResponseStatusException(NOT_FOUND, COMPONENT_ERROR);
+        throw new ResponseStatusException(CREATED, "Resource created successfully");
     }
 
     @DeleteMapping(path = "title={resourceTitle}")
     public void deleteResourceByTitle(@PathVariable("subjectTitle") String title,
                                       @PathVariable("componentType") String type,
                                       @PathVariable("resourceTitle") String resourceTitle) {
-        if(resourceService.deleteResourceByTitle(title, type, resourceTitle) == 0) {
-            throw new IllegalStateException("Resource not found");
-        }
+        if(subjectService.getSubjectByTitle(title).isEmpty())
+            throw new ResponseStatusException(NOT_FOUND, SUBJECT_ERROR);
+        if(componentService.getComponentByType(title, type).isEmpty())
+            throw new ResponseStatusException(NOT_FOUND, COMPONENT_ERROR);
+        if(resourceService.deleteResourceByTitle(title, type, resourceTitle) == 0)
+            throw new ResponseStatusException(NOT_FOUND, RESOURCE_ERROR);
+        throw new ResponseStatusException(NO_CONTENT, "Resource deleted successfully");
     }
 
     @PutMapping(path = "title={resourceTitle}")
@@ -54,8 +83,12 @@ public class ResourceController {
                                       @PathVariable("componentType") String type,
                                       @PathVariable("resourceTitle") String resourceTitle,
                                       @RequestBody Resource resource) {
+        if(subjectService.getSubjectByTitle(title).isEmpty())
+            throw new ResponseStatusException(NOT_FOUND, SUBJECT_ERROR);
+        if(componentService.getComponentByType(title, type).isEmpty())
+            throw new ResponseStatusException(NOT_FOUND, COMPONENT_ERROR);
         if(resourceService.updateResourceByTitle(title, type, resourceTitle, resource) == 0) {
-            throw new IllegalStateException("Resource not found");
+            throw new ResponseStatusException(NOT_FOUND, RESOURCE_ERROR);
         }
     }
 }
